+++ spec/helpers/project_helper_spec.rb
@@ -118,4 +118,24 @@ RSpec.describe ProjectHelper, type: :helper do
       end
     end
   end
+
+  describe "#all_conditions_met_tag" do
+    context "when all_conditions_met is true" do
+      let(:task_list) { create(:voluntary_conversion_task_list, conditions_met_confirm_all_conditions_met: true) }
+      let(:project) { build(:conversion_project, task_list: task_list) }
+
+      it "returns a turquoise tag with 'yes' text" do
+        expect(helper.all_conditions_met_tag(project)).to include("turquoise", "yes")
+      end
+    end
+
+    context "when all_conditions_met is false" do
+      let(:task_list) { create(:voluntary_conversion_task_list, conditions_met_confirm_all_conditions_met: nil) }
+      let(:project) { build(:conversion_project, task_list: task_list) }
+
+      it "returns a blue tag with 'not started' text" do
+        expect(helper.all_conditions_met_tag(project)).to include("blue", "not started")
+      end
+    end
+  end
 end